Update of analogSignal
Parameter types changed from float to uint16.
